+++ src/test/java/com/openggf/game/rewind/TestS1BadnikGenericRecreate.java
@@ -6,8 +6,11 @@ import com.openggf.game.sonic1.objects.badniks.Sonic1BatbrainBadnikInstance;
 import com.openggf.game.sonic1.objects.badniks.Sonic1BombBadnikInstance;
 import com.openggf.game.sonic1.objects.badniks.Sonic1BurrobotBadnikInstance;
 import com.openggf.game.sonic1.objects.badniks.Sonic1BuzzBomberBadnikInstance;
+import com.openggf.game.sonic1.objects.badniks.Sonic1ChopperBadnikInstance;
 import com.openggf.game.sonic1.objects.badniks.Sonic1CrabmeatBadnikInstance;
+import com.openggf.game.sonic1.objects.badniks.Sonic1JawsBadnikInstance;
 import com.openggf.game.sonic1.objects.badniks.Sonic1MotobugBadnikInstance;
+import com.openggf.game.sonic1.objects.badniks.Sonic1NewtronBadnikInstance;
 import com.openggf.game.sonic1.objects.badniks.Sonic1RollerBadnikInstance;
 import com.openggf.game.sonic1.objects.badniks.Sonic1YadrinBadnikInstance;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -24,8 +27,11 @@ class TestS1BadnikGenericRecreate {
             Sonic1BombBadnikInstance.class,
             Sonic1BurrobotBadnikInstance.class,
             Sonic1BuzzBomberBadnikInstance.class,
+            Sonic1ChopperBadnikInstance.class,
             Sonic1CrabmeatBadnikInstance.class,
+            Sonic1JawsBadnikInstance.class,
             Sonic1MotobugBadnikInstance.class,
+            Sonic1NewtronBadnikInstance.class,
             Sonic1RollerBadnikInstance.class,
             Sonic1YadrinBadnikInstance.class
     })